Just need to fine tune collision boxes

+++ src/mainPackage/Pipe.java
@@ -9,7 +9,7 @@ import com.sun.j3d.utils.geometry.GeometryInfo;
 import com.sun.j3d.utils.geometry.NormalGenerator;
 
 public class Pipe extends Shape3D{
-	public Pipe(int y) {
+	public Pipe(double y) {
 		GeometryInfo gi=new GeometryInfo(GeometryInfo.POLYGON_ARRAY);
 		Point3d[] vertices=
 				{
